Move lot tests to separate file
